transaction_server/requirements.txt
asgiref==3.3.1
Django==3.0.5
django-debug-toolbar==3.2
django-filter==2.4.0
django-redis==4.12.1
djangorestframework==3.12.2
djongo==1.3.4
pymongo==3.11.3
pytz==2021.1
redis==3.5.3
sqlparse==0.2.4
